+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Expect active development and potentially significant breaking changes in the `0
 
 - ...
 - Fix bug that caused updateQuery and reducers to run on stopped queries [PR #1054](https://github.com/apollostack/apollo-client/pull/1054)
+- Ensure transporters are using `isomorphic-fetch` instead of `whatwg-fetch` for universal compatibility [PR #1018](https://github.com/apollostack/apollo-client/pull/1018)
 
 ### 0.5.21